Refactor and organize files
